优化性能,, 添加图标, 增加连接树的store. 封装GridTable, TableInfomation
