+++ src/components/Panel/TablePanel.vue
@@ -1,30 +1,26 @@
 <script setup lang="ts">
-  import 'ag-grid-community/styles/ag-grid.css';
-  import 'ag-grid-community/styles/ag-theme-balham.css';
-  import { AgGridVue } from 'ag-grid-vue3';
   import { SimpleNode } from '~/components/ConnectManager/index';
+  import GridTable from '~/components/Base/GridTable.vue';
+  import TableInformation from '~/components/Panel/TableInformation.vue';
+
   import { ref, computed, reactive, PropType, watch, nextTick } from 'vue';
   import Manager from '~/utils/link_manager';
-  import DefaultRenderer from './render.js';
-  import { FIELD_TYPE_MAP } from '~/utils/constants';
-  import { format } from 'sql-formatter';
-  import { useClipboard } from '@vueuse/core';
-  import { IconCopy, IconInfoCircleFill, IconStar, IconThunderbolt } from '@arco-design/web-vue/es/icon';
-  import { Notification } from '@arco-design/web-vue';
-  import { getIconRes } from '~/utils/res';
-  import { TableFields } from './table';
+  import { getImageRes } from '~/utils/res';
   import { useI18n } from 'vue-i18n';
-
-  const manager: Manager = Manager.getInstance();
-  let gridApi = reactive({});
-  let gridColumnApi = reactive({});
+  import { Icon } from '@arco-design/web-vue';
 
   const props = defineProps({
     node: {
       type: Object as PropType<SimpleNode>,
     },
   });
-  console.error('First show', props.node);
+
+  const IconFont = Icon.addFromIconFontCn({
+    src: getImageRes('iconfont/iconfont.js'),
+  });
+  const manager: Manager = Manager.getInstance();
+
+  // console.error('First show', props.node);
 
   const pageInfo = reactive({
     page: 1,
@@ -32,28 +28,7 @@
   });
   const fields = ref([]);
   const rows = ref([]);
-  const tableInfos = ref(null);
-  const createSQL = ref('');
-  const { text, copy, copied, isSupported } = useClipboard({ createSQL });
   const { t } = useI18n();
-
-  const loadTableInfo = async () => {
-    const conn = manager.get(props.node?.meta?.Param.serverKey);
-    console.log('Load table info ' + pageInfo.page);
-
-    const resp = await conn.query(
-      'SHOW FULL FIELDS FROM `' + props.node.meta?.DatabaseName + '`.`' + props.node.title + '`',
-      []
-    );
-    console.log(resp);
-    tableInfos.value = resp.data;
-
-    const tableResp = await conn.query(
-      'SHOW CREATE TABLE `' + props.node.meta?.DatabaseName + '`.`' + props.node.title + '`',
-      []
-    );
-    createSQL.value = format(tableResp.data[0]['Create Table']);
-  };
 
   const loadPage = async () => {
     let start = (pageInfo.page - 1) * pageInfo.page_size;
@@ -62,9 +37,9 @@
 
     const resp = await conn.query(
       'SELECT * FROM ' +
-        props.node.meta?.DatabaseName +
+        props.node?.meta?.DatabaseName +
         '.' +
-        props.node.title +
+        props.node?.title +
         ' LIMIT ' +
         start +
         ',' +
@@ -75,82 +50,18 @@
 
     fields.value = resp.fields;
     rows.value = resp.data;
-    gridApi.setRowData(resp.data);
-    nextTick(() => {
-      autoResize();
-    });
+    // gridApi.setRowData(resp.data);
+    // nextTick(() => {
+    //   autoResize();
+    // });
   };
 
-  const columnDefs = computed(() => {
-    return fields.value.map((item: any, index: any) => {
-      let tip = '';
-      tip += 'Type: ' + FIELD_TYPE_MAP[item.type] ?? item.type;
-      if ((item.flags & 2) != 0) {
-        if ((item.flags & 512) != 0) {
-          tip += '  主键(A)';
-        } else {
-          tip += '  主键';
-        }
-      }
-      let field: Record<string, any> = {
-        headerName: item.name,
-        headerTooltip: tip,
-        field: item.name,
-        width: 90,
-        minWidth: 40,
-        maxWidth: 200,
-        editable: item.name == 'id' ? false : true,
-        suppressMenu: true,
-        sort: null,
-        cellRenderer: 'DefaultRenderer',
-      };
-
-      if (index == 0) {
-        //主键
-        field.sort = 'asc';
-        field.headerCheckboxSelection = true;
-        field.checkboxSelection = true;
-        field.showDisabledCheckboxes = true;
-      }
-      return field;
-    });
-  });
-
-  const gridOptions = reactive({
-    components: {
-      DefaultRenderer: DefaultRenderer,
-    },
-    tooltipShowDelay: 500,
-    defaultColDef: { flex: 1, sortable: false, resizable: true },
-    // debounceVerticalScrollbar: true,
-    suppressRowClickSelection: 'always',
-    rowSelection: 'multiple',
-  });
-
+  loadPage();
   const rowData = computed(() => {
-    // console.warn(rows.value);
     return rows.value.map((item: any) => {
       return item;
     });
   });
-
-  const rowValueChange = () => {
-    console.log(arguments);
-  };
-
-  const autoResize = () => {
-    const allColumnIds = [];
-    gridColumnApi.getColumns().forEach((column: any) => {
-      allColumnIds.push(column.getId());
-    });
-    gridColumnApi.autoSizeColumns(allColumnIds, false);
-  };
-
-  const onGridReady = (params: any) => {
-    gridApi = params.api;
-    gridColumnApi = params.columnApi;
-    autoResize();
-  };
 
   const isNext = computed(() => {
     if (rows.value.length == pageInfo.page_size) {
@@ -174,49 +85,16 @@
     }
   };
 
-  const onSelectionChanged = () => {};
-
-  const copyData = () => {
-    // const selectedNodes = gridApi.getSelectedNodes();
-    // console.log(selectedNodes);
-    const selectedRows = gridApi.getSelectedRows();
-
-    let rows: string[] = [];
-    selectedRows.forEach((element: any) => {
-      let tab = [];
-      for (let k in element) {
-        tab.push(`"${element[k]}"`);
-      }
-      rows.push(tab.join('\t'));
-    });
-    console.log(rows.join('\n'));
-  };
-
-  const exportData = () => {
-    gridApi.exportDataAsCsv();
-  };
-
-  watch(
-    () => props.node,
-    async () => {
-      pageInfo.page = 1;
-      await loadPage();
-    },
-    {
-      immediate: true,
-    }
-  );
-
-  const columns = ref(TableFields);
-
-  const copySQL = () => {
-    copy(createSQL.value);
-    Notification.info({
-      title: '复制成功',
-      content: '',
-    });
-  };
-  loadTableInfo();
+  // watch(
+  //   () => props.node,
+  //   async () => {
+  //     pageInfo.page = 1;
+  //     await loadPage();
+  //   },
+  //   {
+  //     immediate: true,
+  //   }
+  // );
 </script>
 
 <template>
@@ -224,9 +102,9 @@
     <a-tabs size="mini" :position="'top'">
       <a-tab-pane key="data">
         <template #title>
-          <img style="width: 8px; height: 8px" :src="getIconRes('table.png')" />{{
+          <icon-font class="iconfont" type="icon-table" size="14" /><span style="margin-left: 6px">{{
             t('message.tablepanel.data')
-          }}</template
+          }}</span></template
         >
         <div class="toolbar">
           <div></div>
@@ -237,44 +115,19 @@
           </div>
         </div>
 
-        <ag-grid-vue
-          v-contextmenu:contextmenu
-          @grid-ready="onGridReady"
-          @onRowValueChanged="rowValueChange"
-          class="ag-theme-balham container"
-          :gridOptions="gridOptions"
-          :columnDefs="columnDefs"
-          :rowData="rowData"
-          @selection-changed="onSelectionChanged"
-        >
-        </ag-grid-vue>
+        <GridTable :rows="rowData" :fields="fields" />
       </a-tab-pane>
       <a-tab-pane key="info">
-        <template #title> <IconInfoCircleFill />{{ t('message.tablepanel.info') }}</template>
-
-        <div class="container-info">
-          <div style="margin: 5px 15px; user-select: auto">字段信息:</div>
-          <a-table
-            :bordered="{ cell: true }"
-            column-resizable
-            style="margin: 5px 15px; user-select: auto"
-            :pagination="false"
-            size="mini"
-            :columns="columns"
-            :data="tableInfos"
-          >
-            <template #Pk="{ record }">
-              <IconStar style="color: cf1322" v-if="record.Key == 'PRI'" />
-              <IconThunderbolt style="color: #ffc53d" v-if="record.Key == 'MUL'" />
-            </template>
-          </a-table>
-          <div style="margin: 15px 15px 5px 15px; user-select: auto">Table DDL:</div>
-
-          <div class="code">
-            <highlightjs language="sql" :autodetect="false" :code="createSQL" />
-            <div class="copy"><IconCopy size="22" @click="copySQL" /></div>
-          </div>
-        </div>
+        <template #title>
+          <icon-font class="iconfont" type="icon-infomation" size="14" /><span style="margin-left: 6px">{{
+            t('message.tablepanel.info')
+          }}</span></template
+        >
+        <TableInformation
+          :server-key="node?.meta?.Param.serverKey"
+          :database="node?.meta?.DatabaseName"
+          :table="node?.title"
+        />
       </a-tab-pane>
     </a-tabs>
 
@@ -306,40 +159,7 @@
   .panel {
     height: calc(var(--bodyHeight) - 36px);
     width: var(--bodyWidth);
-    .container {
-      width: 100%;
-      height: calc(var(--bodyHeight) - 36px - 28px - 32px);
-    }
-    .container-info {
-      background-color: #fafafa;
-      width: 100%;
-      height: calc(var(--bodyHeight) - 36px - 32px);
-      overflow-y: auto;
-      .code {
-        margin: 0 15px 15px 15px;
-        user-select: text;
 
-        font-size: 13px;
-        position: relative;
-        &:hover {
-          .copy {
-            display: block;
-            color: #424242;
-            &:hover {
-              color: #5975f5;
-            }
-          }
-        }
-        .copy {
-          display: none;
-          position: absolute;
-          cursor: pointer;
-          right: 8px;
-          top: 8px;
-          z-index: 100;
-        }
-      }
-    }
     .toolbar {
       height: 28px;
       line-height: 28px;
